requirements.txt
Flask
Flask-OpenID
Flask-Script
Flask-WTF
Flask-SQLAlchemy
Flask-Bootstrap
SQLAlchemy
psycopg2
alembic
pytz
python-dateutil
pyLibravatar
# pydns
requests
gitpython
redis
asyncio_redis
ipdb
backports.typing
# copr - from system python3-copr
# psycopg2 - from system python3-psycopg2
lxml

tox
pytest
pytest-cov
testing.postgresql

sphinx
sphinxcontrib-httpdomain
